+++ R/Private.R
@@ -1,0 +1,51 @@
+# Copyright 2020 Observational Health Data Sciences and Informatics
+#
+# This file is part of CohortDiagnostics
+# 
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# 
+#     http://www.apache.org/licenses/LICENSE-2.0
+# 
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+createIfNotExist <- function(type, name = c(), recursive = TRUE) {
+  if (!is.null(type)) {
+    if (length(name) == 0) {
+      stop(ParallelLogger::logError("Must specify ", name))
+    }
+    
+    if (type %in% c('folder')) {
+      for (i in (1:length(name))) {
+        if (!file.exists(name)) {
+          dir.create(name, recursive = recursive)
+          ParallelLogger::logInfo("Created ", type, " at ", name)
+        } else {
+          ParallelLogger::logInfo(type, " already exists at ", name)
+        }
+      }
+    }
+  }
+}
+
+
+writeToCsv <- function(data, fileName, incremental = FALSE, ...) {
+  colnames(data) <- SqlRender::camelCaseToSnakeCase(colnames(data))
+  if (incremental) {
+    params <- list(...)
+    names(params) <- SqlRender::camelCaseToSnakeCase(names(params))
+    params$data = data
+    params$fileName = fileName
+    do.call(saveIncremental, params)
+  } else {
+    readr::write_csv(data, fileName)
+  }
+}
+
+