+++ man-roxygen/ErrorMessage.R
@@ -1,0 +1,4 @@
+#' @param errorMessage        An object of type \code{“AssertCollection”} as created using the
+#'                            \code{\link[checkmate]{makeAssertCollection}} function in the
+#'                            checkmate package. If NULL, then new errorMessage object of 
+#'                            \code{“AssertCollection”} type is created to collect assertions.